chore(@clayui/demos): update demos stories
